+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/tests/Azure.Monitor.OpenTelemetry.AspNetCore.Tests/LiveMetrics/DocumentTests/RequestTests.cs
@@ -25,8 +25,11 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore.Tests.LiveMetrics.DocumentTests
         {
         }
 
-        [Fact]
-        public void VerifyRequestAttributes()
+        [Theory]
+        [InlineData("http", "example.com", 8080, "/search", "?q=OpenTelemetry", "http://example.com:8080/search?q=OpenTelemetry")]
+        [InlineData("http", "example.com", 80, "/search", "?q=OpenTelemetry", "http://example.com/search?q=OpenTelemetry")] // as a sideeffect of setting as new Uri, the default port is removed from the Absolute Uri.
+        [InlineData("http", "example.com", 443, "/search", "?q=OpenTelemetry", "http://example.com/search?q=OpenTelemetry")] // we do not record the 443 port.
+        public void VerifyRequestAttributes(string urlScheme, string serverAddress, int serverPort, string urlPath, string urlQuery, string expectedUrl)
         {
             // SETUP
             var uniqueTestId = Guid.NewGuid();
@@ -46,11 +49,22 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore.Tests.LiveMetrics.DocumentTests
             using var requestActivity = activitySource.StartActivity(name: "HelloWorld", kind: ActivityKind.Server);
             Assert.NotNull(requestActivity);
             requestActivity.SetTag("http.response.status_code", 200);
-            requestActivity.SetTag("url.scheme", "http");
-            requestActivity.SetTag("server.address", "example.com");
-            requestActivity.SetTag("server.port", "8080");
-            requestActivity.SetTag("url.path", "/search");
-            requestActivity.SetTag("url.query", "?q=OpenTelemetry");
+            requestActivity.SetTag("url.scheme", urlScheme);
+            requestActivity.SetTag("server.address", serverAddress);
+            requestActivity.SetTag("server.port", serverPort);
+            requestActivity.SetTag("url.path", urlPath);
+            requestActivity.SetTag("url.query", urlQuery);
+            requestActivity.SetTag("customKey1", "customValue1");
+            requestActivity.SetTag("customKey2", "customValue2");
+            requestActivity.SetTag("customKey3", "customValue3");
+            requestActivity.SetTag("customKey4", "customValue4");
+            requestActivity.SetTag("customKey5", "customValue5");
+            requestActivity.SetTag("customKey6", "customValue6");
+            requestActivity.SetTag("customKey7", "customValue7");
+            requestActivity.SetTag("customKey8", "customValue8");
+            requestActivity.SetTag("customKey9", "customValue9");
+            requestActivity.SetTag("customKey10", "customValue10");
+            requestActivity.SetTag("customKey11", "customValue11");
             requestActivity.Stop();
 
             var requestDocument = DocumentHelper.ConvertToRequestDocument(requestActivity);
@@ -59,7 +73,9 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore.Tests.LiveMetrics.DocumentTests
             Assert.Equal(DocumentType.Request, requestDocument.DocumentType);
             Assert.Equal("HelloWorld", requestDocument.Name);
             Assert.Equal("200", requestDocument.ResponseCode);
-            Assert.Equal("http://example.com:8080/search?q=OpenTelemetry", requestDocument.Url.AbsoluteUri);
+            Assert.Equal(expectedUrl, requestDocument.Url.AbsoluteUri);
+
+            VerifyCustomProperties(requestDocument);
 
             // The following "EXTENSION" properties are used to calculate metrics. These are not serialized.
             Assert.Equal(requestActivity.Duration.TotalMilliseconds, requestDocument.Extension_Duration);